U4/requirements.txt
pygamenu==0.0.56
pygame-ce==2.5.5